Add basic app backend/frontend skeleton
The backend is a simple flask app, located in symptoms/symptoms.
The environment is managed through pipenv, with the prod dependencies
specified in setup.cfg.
The app frontend is angularjs 1.6, and located in symptoms/symptoms_js.
Webpack bundles all the js and dumps it into the flask app's static
assets folder, under the gen/ directory.